pyproject.toml
[project]
name = "exotel-chatbot-dial-out"
version = "0.1.0"
description = "Exotel dial-out for Pipecat"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pipecat-ai[webrtc,daily,silero,openai,local-smart-turn-v3,runner,elevenlabs]>=0.0.85",
  "pipecatcloud>=0.2.4",
  "python-dotenv",
  "fastapi",
  "uvicorn",
  "pydantic>=2.0.0",
  "httpx",
  "psycopg2"
]
